+++ src/components/App.jsx
@@ -3,7 +3,7 @@ import Product from './productVariant/Product.jsx'
 import CustomerReviewsContainer from './CustomerReviewsContainer';
 import CWBA from './CustomerWhoBoughtAlsoCmpnt/CustomersWhoBoughtAlsoAll'
 import QAcomponent from './QAcomponent';
-import ComparisonGrid from './ComparisonGrid';
+import ComparisonGrid from './ComparisonGrid/ComparisonGrid.jsx';
 import CImagesAll from './CustomerImageCmpnt/CustomerImagesAll';
 import Appples from './RatingSummary/Appples';
