+++ src/components/ComparisonGrid/ComparisonItem.jsx
@@ -1,0 +1,5 @@
+import React, { Component } from 'react';
+
+export default function ComparisonItem() {
+  return <h1>Hello </h1>;
+}
