+++ tests/comparisongrid.spec.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import ComparisonGrid from '../src/components/ComparisonGrid/ComparisonGrid.jsx';
+
+it("renders without crashing", () => {
+    const div = document.createElement("div");
+    ReactDOM.render(<ComparisonGrid/>, div);
+})
